backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pillow>=10.3.0
numpy>=1.26.0
httpx>=0.27.0
